requirements.txt
Flask
requests
beautifulsoup4
python-dotenv
pytz
pymysql
psycopg2-binary
selenium
urllib3==1.26.16
python-telegram-bot==13.15
APScheduler
